Changed: Refactor task modules to prepare for TaskRunner expansion
- Split task.rs into task/mod.rs (implementation) and task/tests.rs (tests) in both rig and serdesAI adapters
- Reserved task/runner.rs for future TaskRunner implementation
- Preserved public API: TaskTool and TaskArgs exports unchanged
- All 288 tests pass, clippy clean
